refactor: Combine FlatTransformer into PretranslationTransformer, in new transformer.py
To enable testing and remove dependency loops,
a callback argument is added to the transform's constructor.
