+++ pontoon/base/tests/test_fluent.py
@@ -5,7 +5,6 @@ from fluent.syntax import FluentParser, FluentSerializer
 from textwrap import dedent
 
 from pontoon.base.fluent import (
-    FlatTransformer,
     get_simple_preview,
     is_plural_expression,
     create_locale_plural_variants,
@@ -13,8 +12,6 @@ from pontoon.base.fluent import (
 
 
 parser = FluentParser()
-serializer = FluentSerializer()
-transformer = FlatTransformer()
 
 
 MULTILINE_SOURCE = """key =
@@ -75,124 +72,6 @@ SIMPLE_TRANSLATION_TESTS = OrderedDict(
         ("number-literal", ("key = { 1 }", "{ 1 }")),
     )
 )
-
-
-def test_flat_transformer_value_single_element():
-    # Do not modify value with single element
-    source_string = dedent(
-        """
-        title = My Title
-    """
-    )
-
-    res = parser.parse_entry(source_string)
-    transformer.visit(res)
-
-    assert len(res.value.elements) == 1
-    assert res.value.elements[0].value == "My Title"
-
-
-def test_flat_transformer_attribute_single_element():
-    # Do not modify attributes with single element
-    source_string = dedent(
-        """
-        title =
-            .foo = Bar
-    """
-    )
-
-    res = parser.parse_entry(source_string)
-    transformer.visit(res)
-
-    assert len(res.attributes) == 1
-    assert len(res.attributes[0].value.elements) == 1
-    assert res.attributes[0].value.elements[0].value == "Bar"
-
-
-def test_flat_transformer_single_select_expression():
-    # Do not modify value with a single select expression
-    source_string = dedent(
-        """
-        my-entry =
-            { PLATFORM() ->
-                [variant] Hello!
-                *[another-variant] World!
-            }
-    """
-    )
-
-    res = parser.parse_entry(source_string)
-    transformer.visit(res)
-
-    assert len(res.value.elements) == 1
-    assert (
-        res.value.elements[0].expression.variants[0].value.elements[0].value == "Hello!"
-    )
-    assert (
-        res.value.elements[0].expression.variants[1].value.elements[0].value == "World!"
-    )
-
-
-def test_flat_transformer_value_several_elements():
-    # Flatten value with several elements
-    source_string = dedent(
-        """
-        title = My { $awesome } Title
-    """
-    )
-
-    res = parser.parse_entry(source_string)
-    transformer.visit(res)
-
-    assert len(res.value.elements) == 1
-    assert res.value.elements[0].value == "My { $awesome } Title"
-
-
-def test_flat_transformer_attribute_several_elements():
-    # Flatten attribute with several elements
-    source_string = dedent(
-        """
-        title =
-            .foo = Bar { -foo } Baz
-    """
-    )
-
-    res = parser.parse_entry(source_string)
-    transformer.visit(res)
-
-    assert len(res.attributes) == 1
-    assert len(res.attributes[0].value.elements) == 1
-    assert res.attributes[0].value.elements[0].value == "Bar { -foo } Baz"
-
-
-def test_flat_transformer_value_and_attributes():
-    # Flatten value and attributes
-    source_string = dedent(
-        """
-        batman = The { $dark } Knight
-            .weapon = Brain and { -wayne-enterprise }
-            .history = Lost { 2 } parents, has { 1 } "$alfred"
-    """
-    )
-
-    res = parser.parse_entry(source_string)
-    transformer.visit(res)
-
-    assert len(res.value.elements) == 1
-    assert res.value.elements[0].value == "The { $dark } Knight"
-
-    assert len(res.attributes) == 2
-
-    assert len(res.attributes[0].value.elements) == 1
-    assert (
-        res.attributes[0].value.elements[0].value == "Brain and { -wayne-enterprise }"
-    )
-
-    assert len(res.attributes[1].value.elements) == 1
-    assert (
-        res.attributes[1].value.elements[0].value
-        == 'Lost { 2 } parents, has { 1 } "$alfred"'
-    )
 
 
 @pytest.mark.parametrize("k", SIMPLE_TRANSLATION_TESTS)
@@ -338,6 +217,7 @@ def test_create_locale_plural_variants(locale_a):
     """
     )
 
+    serializer = FluentSerializer()
     assert serializer.serialize_entry(message) == serializer.serialize_entry(
         parser.parse_entry(expected)
     )
